+++ tournament.js
@@ -12,17 +12,4 @@ var t = {
   robin: algs.robin
 };
 
-// A general parser that will find out which tournament type to parse it as
-t.parse = function (str) {
-  var data = JSON.parse(str);
-  if (data.type == null) {
-    // NB: First time parsing of old ones need to go to the specific class
-    throw new Error("Need to migrate over pre1.0 serialized tournaments");
-  }
-  if (typeof t[data.type] !== 'function') {
-    throw new Error("Invalid tournament type: " + data.type);
-  }
-  return t[data.type].parse(data);
-};
-
 module.exports = t;